% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneDsample.R
\name{oneDsample}
\alias{oneDsample}
\title{Single variable Rejection Sampling}
\usage{
oneDsample(f, N, lb = -Inf, ub = Inf, discrete = FALSE)
}
\arguments{
\item{f}{the pdf that we are sampling from}

\item{N}{the number of attempted samples.}

\item{lb}{lower bound of support of f}

\item{ub}{upper bound of support of f}
}
\value{
A vector containing samples from pdf
}
\description{
This function implements single variabel rejection sampling for rvs with bounded support and which have have bounded pdf.
}
\examples{


f <- function(x) {ifelse(0 < x & x < 1, 4*x^3, 0)}
a <- oneDsample(f,10000)
ggplot(a,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

f <- function(x) {1/pi/(1+x^2)}
a <- oneDsample(f,20000)
ggplot(a,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

f <- function(x) {dunif(x, -100, 100)}
a <- oneDsample(f,50000, -100, 100)
a <- oneDsample(f,500)
ggplot(a,aes(x)) + geom_density() + stat_function(fun = f, color = "red")

}
