% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TwoDsample.R
\name{twoDsample}
\alias{twoDsample}
\title{Conditional variable Rejection Sampling}
\usage{
twoDsample(f, N, lbx = -5000, ubx = 5000, lby = -5000, uby = 5000)
}
\arguments{
\item{f}{the joint pdf that we are sampling from}

\item{N}{the number of attempted samples.}

\item{lbx}{lower bound of support x of f}

\item{ubx}{upper bound of support x of f}

\item{lby}{upper bound of support y of f}

\item{uby}{upper bound of support y of f}
}
\value{
A vector containing samples from pdf
}
\description{
This function implements conditional variabel rejection sampling for rvs with bounded support x,y and which have have bounded pdf.
}
\examples{

jointPFF <- function(x){
x1 = x[1]
x2 = x[2]
ifelse(0<x1 & x1<1 & 0<x2 & x2<1 & 0<x1+x2 & x1+x2<1, 24*x1*x2, 0)}

f <- function(x){
x1 = x[1]
x2 = x[2]
ifelse(x2>0, 1/pi/(1+x1^2) * 0.05*exp(-0.05*x2), 0)}

a <- twoDsample(f = f, N=10000)
ggplot(a, aes(x, y)) +  geom_density_2d()
}
